lightweight NCriticalPathModel file by having separate file NCriticalPathModelLoader
